rh: restructuring
